third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMorphologyElement-svgdom-radius-call-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'radius' property of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.radiusX.baseVal is 6
PASS morphologyElement.radiusY.baseVal is 6
PASS morphologyElement.radiusX.baseVal is 4
PASS morphologyElement.radiusY.baseVal is 4
PASS successfullyParsed is true

TEST COMPLETE
